Added Course page (#3)
During the project we will need a template page for all the tasks, tasks which will have generated content for each page.

This commit adds the Course HTML page, which is the template needed, implements the link between it and the other pages and adds the necessary classes in the style.css file.